agent-blueprint/agentcore-runtime-a2a-stack/report-writer/src/requirements.txt
# Strands Agents with A2A support (latest version)
strands-agents[a2a]>=1.16.0

# AWS SDK (latest versions)
boto3>=1.40.74
botocore>=1.40.74

# AgentCore SDK (latest version)
bedrock-agentcore>=1.0.6

# Document processing
python-docx>=1.1.2

# Web framework (latest versions)
fastapi>=0.121.2
uvicorn>=0.38.0

# Utilities
python-dateutil>=2.9.0
